I tried to fix the problems with my last commit,
i wrote a new makefile (makefile_new) but it gives me failures that i dont understand
because in my opinion there is no syntax error, it compiles with my ide
but not with the makefile im writing.

i deleted fibonacci.h and .cpp and put the code all in fibonacci_test.cpp.

Also added the included files from the ide makefile maybe then it is atleast possible to compile then.